Stop a typed-in storage quota from 500ing when a client is created
Filling the "Storage quota (MB)" field on the new-client form raised a
TypeError and the request died with a 500. Leaving it blank worked, which
is why it reached a release: that path goes through `null ?? 0`, and the
0 is an int.

The `integer` validation rule checks that a value looks like an integer.
It does not convert it. `$request->validate()` returns the raw input, so
the form field arrives as the string "2048" -- and the create form types
that field as a string in React, so it is a string even over JSON. Both
controllers declare strict_types, so handing it to
`ClientAccounts::create()`'s `int $storageQuotaMb` is a TypeError.

Fixed on both surfaces that call create(): the staff screen and
/api/v1/clients. The API twin had the same defect, reachable by sending
the quota as a quoted JSON value or a form-encoded body -- its own create
test only ever sent a JSON number.

Two more call sites had the same shape and are cast too, though nothing
sends them a string today: the share-link download cap and a comment's
reply_to. Both are safe only because a frontend file happens to call
Number() first, which is a fact about that file rather than anything the
signature guarantees. The null in each is preserved rather than collapsed
to 0 -- "no cap" is not a cap of zero.

`storage_quota_mb` is also cast on User and Invitation. The column is an
unsignedInteger and both docblocks already promise int; it is read
straight into provision()'s typed parameter when an invitation is
redeemed, and which type a driver hands back is not something that call
site should depend on.

Found on the new files-test rehearsal instance, on its first real use,
against the same build the whole fleet is running.

+++ app/Modules/Clients/Http/Controllers/Api/ClientsController.php
@@ -144,7 +144,12 @@ class ClientsController extends Controller
             name: $validated['name'],
             email: $validated['email'],
             password: $validated['password'],
-            storageQuotaMb: $validated['storage_quota_mb'] ?? 0,
+            // As on the staff screen, and for the same reason: the
+            // `integer` rule accepts a numeric string and does not convert
+            // it. A JSON number arrives as an int and was fine; a
+            // form-encoded body or a quoted JSON value is a string, and
+            // this file is strict_types.
+            storageQuotaMb: (int) ($validated['storage_quota_mb'] ?? 0),
             welcome: false,
         );
